+++ forest/cli/Cargo.toml
@@ -51,18 +51,13 @@ log.workspace = true
 multibase.workspace = true
 nom = "7.1.3"
 num.workspace = true
-once_cell.workspace = true
 pretty_assertions = "1.3.0"
 rpassword.workspace = true
-rust_decimal = "1.28.1"
-rust_decimal_macros = "1.28.1"
 rustyline = "10.1.1"
 serde = { workspace = true, features = ["derive"] }
 serde_json.workspace = true
 serde_tuple.workspace = true
-si-scale = "0.2.1"
 strfmt = "0.2.2"
-strum = "0.24.1"
 tempfile.workspace = true
 ticker = "0.1"
 tokio = { workspace = true, features = ["sync"] }
